Moving messages concern to another module
