+++ app/models/user.rb
@@ -107,9 +107,6 @@ class User < ActiveRecord::Base
     assignment.send_question! question
   end
 
-
-
-
   def self.import_from_json!(body)
     body[:name] = "#{body.delete(:first_name)} #{body.delete(:last_name)}"
     User.where(uid: body[:uid]).update_or_create!(body.except(:id))
